Released 0.9.8.0
git-svn-id: https://slrn.svn.sourceforge.net/svnroot/slrn/trunk@153 b66dfc27-5e3b-0410-ac45-e82dab92a7d9

+++ src/slrnpull.c
@@ -2780,7 +2780,7 @@ static int create_overview_for_dir (Active_Group_Type *g, unsigned int *nums, un
 	  {
 	     bmin = bmax = nums[i];
 	  }
-	else if ((bmin != -1) && (ret || (nums[i] != bmax+1)))
+	else if ((bmin != -1) && (ret || ((int)nums[i] != bmax+1)))
 	  {
 	     headers = slrn_ranges_add (headers, bmin, bmax);
 	     if (ret)
